refactor: rename friendshipRepository to friendshipsRepository
