+++ server/src/modules/friendship/friendship.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { FriendshipRepository } from './repositories/friendship.repository';
-import { PrismaFriendshipRepository } from './infra/db/repositories/prisma-friendship.repository';
+import { FriendshipsRepository } from './repositories/friendships.repository';
+import { PrismaFriendshipsRepository } from './infra/db/repositories/prisma-friendships.repository';
 import { AddFriendService } from './services/add-friend.service';
 import { GetFriendsService } from './services/get-friends.service';
 import { AddFriendshipController } from './infra/http/controllers/add-friendship.controller';
@@ -13,8 +13,8 @@ import { UserModule } from '@modules/user/user.module';
   providers: [
     AddFriendService,
     GetFriendsService,
-    { provide: FriendshipRepository, useClass: PrismaFriendshipRepository },
+    { provide: FriendshipsRepository, useClass: PrismaFriendshipsRepository },
   ],
-  exports: [AddFriendService, GetFriendsService, FriendshipRepository],
+  exports: [AddFriendService, GetFriendsService, FriendshipsRepository],
 })
 export class FriendshipModule {}
