+++ server/src/modules/friendship/repositories/friendships.repository.ts
@@ -1,6 +1,6 @@
 import { Friendship } from '../infra/db/entities/friendship';
 
-export abstract class FriendshipRepository {
+export abstract class FriendshipsRepository {
   abstract create(friendship: Friendship): Promise<void>;
   abstract delete(userId: string, friendId: string): Promise<void>;
   abstract findAllOfUser(userId: string): Promise<Friendship[]>;
