+++ server/src/modules/friendship/services/get-friends.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { FriendshipRepository } from '../repositories/friendship.repository';
+import { FriendshipsRepository } from '../repositories/friendships.repository';
 import { User } from '@modules/user/infra/db/entities/user';
 
 interface IResponse {
@@ -8,11 +8,11 @@ interface IResponse {
 
 @Injectable()
 export class GetFriendsService {
-  constructor(private friendshipRepository: FriendshipRepository) {}
+  constructor(private friendshipsRepository: FriendshipsRepository) {}
 
   async execute(userId: string): Promise<IResponse> {
     // Find all friendship of user
-    const friendships = await this.friendshipRepository.findAllOfUser(userId);
+    const friendships = await this.friendshipsRepository.findAllOfUser(userId);
 
     // Format user friendship items
     const friends = friendships.map((friendship) => {
